+++ src/cli/index.ts
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 import { Command } from "commander";
 import { registerGenerateCommand } from "./commands/generate";
+import { registerAddCommand } from "./commands/add";
 
 const program = new Command();
 
@@ -10,5 +11,6 @@ program
   .description("Enterprise Playwright Automation Platform Generator");
 
 registerGenerateCommand(program);
+registerAddCommand(program);
 
 program.parse(process.argv);